Porting: Video files renamed and C++ed
Video related files are renamed.
Changed indention.
Fixed code style.
