+++ src/minimap.cc
@@ -26,10 +26,10 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/gfx.h"
   #include "src/game.h"
   #include "src/data.h"
 END_EXT_C
+#include "src/gfx.h"
 #include "src/interface.h"
 #include "src/viewport.h"
