+++ src/popup.cc
@@ -26,13 +26,14 @@
 #include <cstring>
 #include <algorithm>
 
+#include "src/misc.h"
 BEGIN_EXT_C
   #include "src/data.h"
   #include "src/game.h"
   #include "src/audio.h"
   #include "src/debug.h"
-  #include "src/gfx.h"
 END_EXT_C
+#include "src/gfx.h"
 #include "src/interface.h"
 #include "src/event_loop.h"
 #include "src/minimap.h"
