+++ src/viewport.cc
@@ -27,12 +27,12 @@
 #include "src/misc.h"
 BEGIN_EXT_C
   #include "src/data.h"
-  #include "src/gfx.h"
   #include "src/game.h"
   #include "src/log.h"
   #include "src/debug.h"
   #include "src/audio.h"
 END_EXT_C
+#include "src/gfx.h"
 #include "src/interface.h"
 #include "src/popup.h"
 #include "src/pathfinder.h"
